change base enum type from byte to int

+++ WindowStyle.cs
@@ -1,9 +1,9 @@
 ﻿namespace KernelTerminal
 {
-    public enum WindowStyle : byte
+    public enum WindowStyle : int
     {
         Default = 0,
         ButtonsHidden = 1 << 0,
-        ProcessHidden= 1 << 1,
+        ProcessHidden = 1 << 1,
     }
 }
